Backendd/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-CORS
Flask-Dance
Flask-Mail
python-dotenv
psycopg2-binary
Authlib

